🚀 GoRouter 구조 개선 및 GlobalShell 도입
Why:
- 기존 `router.dart`, `router_kiosk.dart`, `router_setup.dart` 파일이 각각 관리되며 중복된 코드가 많았음.
- `KioskShell`, `ContentsShell`을 `GlobalShell`로 통합하여 보다 일관된 레이아웃을 유지할 필요가 있음.
- `GoRouter` 설정이 분산되어 있어 유지보수성이 떨어짐.

What:
- `GlobalShell`을 생성하여 모든 `GoRouter` 경로를 통합.
- `kiosk_shell.dart` 내 `KioskShell`, `ContentsShell`을 하나의 `Scaffold`로 통합.
- `router.dart`, `router_kiosk.dart`, `router_setup.dart`를 `router.dart` 및 `go_router.dart`로 정리.
- `GoRouter`의 `ShellRoute`를 활용하여 `GlobalShell`, `ImageShellRouteData`를 정의하고 `SetupMainRouteData`, `KioskRouteData` 등을 포함.
- `router_kiosk.dart` 및 `router_setup.dart` 삭제 및 `router.dart`에서 관리하도록 변경.

How:
- `GlobalShell`을 통해 모든 페이지가 동일한 `Scaffold`를 공유하도록 구조 변경.
- `KioskShell` 내부 로직을 단순화하고, `LoaderOverlay` 적용 부분을 `child`의 `Column` 내부로 이동.
- `GoRouter`에서 `TypedShellRoute`를 활용하여 `GlobalShellRouteData`, `ImageShellRouteData` 추가.
- `router.dart`에서 모든 `GoRouteData` 클래스를 관리하도록 수정.

Tags: #refactor #router #GoRouter #shell-route #ui-improvement
