+++ lib/routers/router.dart
@@ -1,34 +1,165 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:flutter_snaptag_kiosk/features/move_me/screens/global_shell.dart';
 import 'package:flutter_snaptag_kiosk/lib.dart';
 import 'package:go_router/go_router.dart';
 import 'package:material_components/material_screen.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
 
 part 'router.g.dart';
-part 'router_kiosk.dart';
-part 'router_setup.dart';
 
 final GlobalKey<NavigatorState> rootNavigatorKey = GlobalKey<NavigatorState>();
 
-@riverpod
-GoRouter router(Ref ref) {
-  return GoRouter(
-    navigatorKey: rootNavigatorKey,
-    initialLocation: SetupMainRouteData().location,
-    routes: $appRoutes,
-    debugLogDiagnostics: true,
-  );
-}
+final GlobalKey<NavigatorState> globalNavigatorKey = GlobalKey<NavigatorState>(debugLabel: 'global');
 
-@TypedGoRoute<SplashRoute>(
-  path: '/',
+@TypedShellRoute<GlobalShellRouteData>(
+  routes: [
+    TypedGoRoute<SetupMainRouteData>(
+      path: '/setup',
+      routes: [
+        TypedGoRoute<KioskInfoRouteData>(path: 'kiosk-info'),
+        TypedGoRoute<PaymentHistoryRouteData>(path: 'payment-history'),
+        TypedGoRoute<UnitTestRouteData>(path: 'unit-test'),
+        TypedGoRoute<KioskComponentsRouteData>(path: 'kiosk-components'),
+      ],
+    ),
+    TypedGoRoute<KioskRouteData>(
+      path: '/kiosk',
+      routes: [
+        TypedShellRoute<ImageShellRouteData>(
+          routes: <TypedRoute<RouteData>>[
+            TypedGoRoute<PhotoCardUploadRouteData>(path: 'qr'),
+            TypedGoRoute<CodeVerificationRouteData>(path: 'code-verification'),
+            TypedGoRoute<PhotoCardPreviewRouteData>(path: 'preview'),
+            TypedGoRoute<PrintProcessRouteData>(path: 'print-process'),
+          ],
+        ),
+      ],
+    )
+  ],
 )
-class SplashRoute extends GoRouteData {
-  const SplashRoute();
+class GlobalShellRouteData extends ShellRouteData {
+  const GlobalShellRouteData();
 
   @override
-  Widget build(BuildContext context, GoRouterState state) {
-    return const SplashScreen();
+  Widget builder(BuildContext context, GoRouterState state, Widget navigator) {
+    return GlobalShell(child: navigator);
+  }
+}
+
+class ImageShellRouteData extends ShellRouteData {
+  const ImageShellRouteData();
+
+  @override
+  Widget builder(BuildContext context, GoRouterState state, Widget navigator) {
+    return KioskShell(child: navigator);
+  }
+}
+
+class SetupMainRouteData extends GoRouteData {
+  const SetupMainRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: SetupMainScreen(),
+    );
+  }
+}
+
+class PaymentHistoryRouteData extends GoRouteData {
+  const PaymentHistoryRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: PaymentHistoryScreen(),
+    );
+  }
+}
+
+class UnitTestRouteData extends GoRouteData {
+  const UnitTestRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: const UnitTestScreen(),
+    );
+  }
+}
+
+class KioskComponentsRouteData extends GoRouteData {
+  const KioskComponentsRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: const KioskComponentsScreen(),
+    );
+  }
+}
+
+class KioskInfoRouteData extends GoRouteData {
+  const KioskInfoRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: const KioskInfoScreen(),
+    );
+  }
+}
+
+class KioskRouteData extends GoRouteData {
+  const KioskRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: SizedBox(),
+    );
+  }
+}
+
+class PhotoCardUploadRouteData extends GoRouteData {
+  const PhotoCardUploadRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: PhotoCardUploadScreen(),
+    );
+  }
+}
+
+class CodeVerificationRouteData extends GoRouteData {
+  const CodeVerificationRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: CodeVerificationScreen(),
+    );
+  }
+}
+
+class PhotoCardPreviewRouteData extends GoRouteData {
+  const PhotoCardPreviewRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: const PhotoCardPreviewScreen(),
+    );
+  }
+}
+
+class PrintProcessRouteData extends GoRouteData {
+  const PrintProcessRouteData();
+
+  @override
+  Page<void> buildPage(BuildContext context, GoRouterState state) {
+    return NoTransitionPage(
+      child: const PrintProcessScreen(),
+    );
   }
 }
